requirements.txt
Flask==2.3.2
gunicorn==21.2.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
Sphinx==7.2.6
pytest==7.4.0
pytest-flask==1.2.0
